Expose creating agent components for ASP.NET Full Framework
This commit exposes a CreateAgentComponents method on
ElasticApmModule that allows a user to create agent components
for ASP.NET Full Framework and set up the APM agent, before
the ElasticApmModule Init() runs. This is useful in scenarios
where a user wishes to capture transactions/spans or setup filters at application
start.

The CreateAgentComponents returns a derived FullFrameworkAgentComponents
type that SafeAgentSetup method uses to determine whether an already initialized
agent is likely to lead to unexpected behaviour e.g. agent component implementations
not designed to work with ASP.NET Full Framework.

An example usage of CreateAgentComponents is

    public class MvcApplication : HttpApplication
    {
        protected void Application_Start()
        {
            // other application startup e.g. RouteConfig, etc.

            // configure agent filter on startup
            var agentComponents = ElasticApmModule.CreateAgentComponents();
            Agent.Setup(agentComponents);
            Agent.AddFilter((ITransaction t) =>
            {
                t.SetLabel("foo", "bar");
                return t;
            });
        }
    }

Fixes #1098
Fixes #1045
